requirement.txt
google-adk
streamlit
nest-asyncio
litellm